% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{partition.functions}
\alias{partition.functions}
\title{Data: List of built-in (non-exported) partitioning functions}
\format{A character vector}
\usage{
partition.functions
}
\description{
Data: List of built-in (non-exported) partitioning functions
}
\keyword{datasets}

