% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{partition.methods}
\alias{partition.methods}
\title{Data: List of supported partitioning methods}
\format{A character vector}
\usage{
partition.methods
}
\description{
Data: List of supported partitioning methods
}
\keyword{datasets}

