% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/partitionComponents.R
\name{partition}
\alias{partition}
\title{Partition a component matrix}
\usage{
partition(x, FUN, ...)

partition.functions
# c("fixed", "ann")
}
\arguments{
\item{x}{A numeric matrix}

\item{FUN}{Function to use to determine threshold values for partitioning x}

\item{...}{Optional arguments to FUN}
}
\value{
A matrix comprising values of 1, 0, and -1 to indicate column-wise
  partitioning of the input matrix into positive, neither, and negative sets,
  respectively.
}
\description{
Given a numeric matrix, returns a matrix with the same dimensions, row and
column names but with only 1, 0, and -1 as values.  These values indicate the
partitioning of each column of the matrix into positive, neither, and
negative subsets.
}
\details{
Built-in partition functions are listed in the
  \code{partition.functions} vector.  The ("\code{fixed}") function uses a
  fixed value (specified by an additional parameter, \code{t}) to create
  partitions, while the ("\code{ann}") function uses an artificial neural
  network model to create custom partition thresholds for each component.
  Additional user-defined functions are allowed.  These should return a
  matrix with two rows and as many columns as in \code{x}; the first row of
  this matrix specifies the upper partition threshold for each column of
  \code{x}, while the second row specifies the lower partition threshold.
}
\examples{
x = matrix(rnorm(9), 3, 3)
partition(x, FUN = ".fixed", t = 0.5)
}

